Adding new files for hiopMatrixDense
Broke defs and decls into new file
FYI I should have added this to the previous commit
