+++ src/Drivers/nlpMDS_ex5.hpp
@@ -5,7 +5,7 @@
 
 //this include is not needed in general
 //we use hiopMatrixDense in this particular example for convienience
-#include "hiopMatrix.hpp" 
+#include "hiopMatrixDense.hpp" 
 
 #ifdef HIOP_USE_MPI
 #include "mpi.h"
